backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.6,<3.0
python-multipart>=0.0.20,<1.0
httpx>=0.28,<1.0
duckdb>=1.1,<2.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
scikit-learn>=1.5,<2.0
pyod>=2.0,<3.0
